+++ NEWS.md
@@ -1,6 +1,6 @@
 ## wingen 2.2.0
 
--   New `winkrige_gd()` for improved kriging performance.
+-   New `wkrig_gd()` for improved kriging performance.
 -   Superseded `krig_gd()`: This legacy function now gives a deprecation warning and requires manual installation of the {automap} package.
 -   Minor documentation updates.